Workaround #2668 - Compress facts so that the request size limit triggers less often
This is not the right fix, but more a hackish workaround.

Since 0.25, the facts are transmitted as GET parameters when a
node asks for a catalog. Most proxies or webserver have a size limit
which is sometimes reached. In this case the request is denied
and the node can't get its catalog.

The idea is to compress facts (some non-scientific studies show a
57% fact size decrease for an average node) when transmitting
those when asking for a catalog.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>