Bug #1742 Invalid params to --color outputs 'nil'
This patches fixes a bug where setting an invalid option for "--color"
caused the word "nil" to be printed on every line of the log, instead
of printing out log messages.
Invalid color options now just produce uncolored output.
It seems to me that this isn't important enough to issue a warning
about an invalid setting.

Signed-off-by: Jesse Wolfe <jes5199@gmail.com>