conf/redhat: Consistently pass pidfile option to daemon, killproc, and status
In Red Hat bug #531116 Ruben Kerkhof noted that we weren't using the
--pidfile $pidfile option to daemon.  This caused 'service puppet start'
to fail if puppet had already been started.  To be consistent, we now
pass the pidfile option to daemon, killproc, and status when those
functions provide such an option.  And we only test for the availability
of the -p/--pidfile option in one place.