Fixing #2631 show_diff sometimes doesn't happen
This patch fixes a code-path in Puppet::Type::File::Content#insync? that
was failing to show a diff when the content's source came from a
template.

I was only able to reproduce one of the two such failures that Malcolm
Howe reported, but I think it's likely that they were both caused by
this same bug.

Signed-off-by: Jesse Wolfe <jes5199@gmail.com>