Fix #2797 - tags are not inherited by recursed file sub child
This problem affects all types that generate sub-resources at
evaluation or pre-transaction time.
Thus it is fixed in the transaction, where we make sure we assign
all parent tags to the generated child resource.

A more correct fix would be to make sure we copy the whole context
(ie file, line, version, tags). This fix is planned for Rowlf.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>