Fix for ticket #2844 (file recursion generated vs. explicit prefix)
The routine which was determining if one path was a prefix of another
in arbitrating between explicit and generated resources was using the
raw string for the test without regard to path segments and thus could
be fooled by pairs such as "/tmp/foo" vs. "/tmp/foo2"

Fix was to be path delimiter aware and add a test.

Signed-off-by: Markus Roberts <Markus@reality.com>