Fix for #2869 (SELinux tests failing under some load orderings)
The problem was that the setup for the tests was expecting the defaultprovider
for Selboolean and Selmodule to be called, rather than stubbing it.  This
worked as long as no other spec which initializes the providers was run before
it.

The fix here (stubbing rather than expecting) is minimal but not ideal; if
there were some other provider for these types it could result in a test
indetermenacy (different results depending on which provider) but I'm not
seeing an easy way to address that.

Signed-off-by: Markus Roberts <Markus@reality.com>