Fix for #2657 (retain old setting if config has syntax error)
This appears to be regression introduced by threading changes.  The fix was
to rearrange things to keep the old behaviour (don't clear the settings
until you know the config file parses) and the new (don't nest calls to
synchronize) by:

1. Splitting clear into two parts--clear, which works as before, and
unsafe_clear which it calls and which expects synchronization to be
handled externally.

2. Rearranging the code to recover the previous calling order

3. Trapping syntax errors and turning them into logged messages and a
no-op effect.

4. Fixing reparse to not wrap a call to this code with a synchronize.

5. Tests.

Signed-off-by: Markus Roberts <Markus@reality.com>