Fix for #2881 (ralsh doesn't find individual parsed resources)
Rather than just creating a new featureless resource for cases like:

    ralsh host localhost.localdomain

(and the analogous cases with other parsed resources such as ssh keys, etc),
we first check to see if the appropriate provider can find a resource with
the requested name.  If so, we use it; if not we fall back to the featureless
form.

There are a number of potential shortcomings with this fix:

* It may not be particularly performant if there are many resources to
  be parsed.
* It always (and only) checks the first provider; perhaps it should try
  all providers of the appropriate type.
* It only checks the name property, and not any aliases or host_aliases.

Signed-off-by: Markus Roberts <Markus@reality.com>