Additional fix for #2994 (followed symlinks do not have checksums)
The first patch for #2994, to which this is an extension, exposed
the fact that checksums were not being included in the metadata
for followed links; checksums are needed for managing the contents
of files that are represented on the server as links (links => follow).

This patch adds checksums for followed links and tests to confirm that
it works as expected.