Fix for #2999 (absent package handling on solaris)
The fix for #2940 resulted in a behavior change that could be seen as a
bug.  This was consistent with many of the other providers (see #3000)
but was not the desired behavior.

This patch enhances the patch for #2940 to return {:ensure => :absent} as
the present state in the case that the failure was due to the package not
being installed (as opposed to a source/index failure, etc.).  The other
possibility would have been {:ensure => :purged}.