Fixing #2914 - pre/post hooks now work for transactions
This was built to be used with etckeeper to version control
files in /etc, but can be used for essentially anything.

This patch was built to be added to 0.25.4, so it's a least-modify
approach.  A better approach would be to refactor application/puppet.rb
just a bit so it uses Configurer more.

This is a simple patch - it just defines 'prerun_command' and 'postrun_command'
settings, and runs the appropriate command around each transaction
if they're set.

Signed-off-by: Luke Kanies <luke@reductivelabs.com>