Minimal fix for #3001 (failing to fetch metadata on dangling symlink)
FileTest.exists? returns false if the target of a symlink is missing; in such
cases we still want to continue if the resource is a symlink, as we may be
managing a dangling symlink.  Continuing in such case either gives the desired
behavior or a more specific/informative error message.