Fix for #3094 (libdir should take ":" delimited path)
Actually, File::PATH_SEPARATOR, which is generally, but not always, ":").

Since libdir is also the default for the plugin handler, users will need to
specify it explicitly if a multipart libdir is given (and it will need to be
one of the segments given in the libdir for the plugins to be found).