Partial reversion of patch for #3088 to fix #3104 (Exception misreported)
In my patch for #3088 I made a erroneous assumption about the ruby exception
hierarchy and thus missed the fact that Timeout::error descends from both
SignalError and Interrupt.  This is a partial reversion of the patch for #3088
to let these through so that more useful error messages can be produced.