Fix #3155 - prevent error when using two matching regex in cascade
The following manifest:
case $var {
  /match/: {
     if $var =~ /matchagain/ {
     }
  }
}

is failing because the "=~" operators when matching sets an ephemeral
variable in the scope. But the case regex also did it, and since they
both belong to the same scope, and Puppet variables are immutables, the
scope raises an error.

This patch fixes this issue by adding to the current scope a stack
of ephemeral symbol tables. Each new match operator or case/selector
with regex adds a new scope. When we get out of the case/if/selector
structure the scope is reset to the ephemeral level we were when
entering it.

This way the following manifest produces the correct output:
case $var {
  /match(rematch)/: {
     notice("1. \$0 = $0, \$1 = $1")
     if $var =~ /matchagain/ {
         notice("2. \$0 = $0, \$1 = $1")
     }
     notice("3. \$0 = $0, \$1 = $1")
  }
}
notice("4. \$0 = $0")

And the output is:
1. $0 = match, $1 = rematch
2. $0 = matchagain, $1 = rematch
3. $0 = match, $1 = rematch
4. $0 =

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>