Fix #3229 - use original value in case/selector regex matching
The issue is that case/selectors are downcasing the value before it
is compared to the options.
Unfortunately regex are matching in a case sensitive way, which would
make the following manifest fail:

$var = "CaseSensitive"
case $var {
  /CaseSensitive/: {
     notice("worked")
  }
  default: {
    fail "miserably"
  }
}

This patch fixes the issue by making sure the regexp match is done
one the original (not downcased) value, but still doing a case
sensitive match.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>