Functions are added to a module instead of Scope
We were previously adding them directly to Scope, but
now they're in a module that Scope includes.

This is the first half of #1175 - we can now maintain
environment-specific collections of functions.  We need
some way of tracking which environment a given function
is loaded from.

Well, maybe it's the first third - the core functions
probably need to be added to all of these modules,
or there needs to be a 'common' module that is included by
all of them.

Signed-off-by: Luke Kanies <luke@reductivelabs.com>