Resolving conflicts with ???
When checksum is a parameter corresponding to checksum_type in the filesets
rather than a checksum "strategy" the value :none should generally be computed
rather than provided.  If neither a source nor contents are provided, there is
no need for a checksum and thus it should be :none; otherwise, use the provided
value (or, if it's nil, let it pass through and be replaced by the default
(:md5) in the bowels of the beast).