Fixed #2568 - Add database option 'dbconnections'
This sets the ActiveRecords connection pool size, when connecting to
remote databases (mysql, postgres).  default is 0; the 'pool' argument
is only passed to ActiveRecords when the value is 1 or greater.

Note: This was cherry-picked from
c9f40be6c567d8de328b9d79dde357672323925a from upstream/master. This
fix is currently targetted for rowlf. There was one hunk that touched
a part of the code that is not in 0.25, which is adding support for
'oracle_enhanced' It is a trivial hunk, related to code that is not in
0.25, so I removed it to complete the merge.