Removing any mentions of :casesensitive setting
It is a setting that was added years ago as a backward
compatibility option and even if it still works, which
is questionable, it has no purpose any longer.

It just complicated the code and didn't do much, so it's gone
now.

Also simplified the interface of Leaf#evaluate_match, since it
was now using none of the passed-in options.

Finally, removed/migrated the last of the Selector/CaseStatement
test/unit tests.

Signed-off-by: Luke Kanies <luke@puppetlabs.com>