Removing invalid File integration test
The point of this test was to confirm that we could take
a given file resource and have it work multiple times in
memory, but that's not actually possible given our current
code.  We copy the values over from the source metadata,
and suddenly have no method for differentiation between
values set by the user and those set by the remote source.

Signed-off-by: Luke Kanies <luke@puppetlabs.com>