Fix for #3556 Plussignment value melding
The plussignment operator was constructing the new parameter value by
modifying the param object's value in place (so as to preserve the file
and line information for debugging).  However, when multiple resources
are overridden by the same plussignment this would result in all of the
resources sharing the same value (the union of all the prior values and
the new value), which is wrong.

Instead, we need to give each resource its own copy of the value (e.g.,
a copy of the param object), which this patch implements.

Signed-off-by: Markus Roberts <Markus@reality.com>