Update Red Hat spec file for 0.25.5
With RHEL-6 just around the corner, now is also a good time to adjust
the ruby(selinux) conditional to work there.  The rundir-perms patch
from tarball is applied rather than including it separately in the SRPM.
It makes for one less file to track in downstream distros.  Other minor
changes are pulled in from the Fedora/EPEL spec file.