Fixes #3745 Catch unhandled exception in ssh_authorized_key provider
If the target is not specified it is automatically set to the user's
home directory.  If the user does not exist when the generation of
the target path occurs then an ArgumentError exception is raised
but not caught.  This patch catches the ArgumentError and raises
a Puppet::Error instead to more gracefully notify the user and allow
any remaining resources to be applied.

Signed-off-by: Sean Millichamp <sean@bruenor.org>