Minimal footprint fix for #3751 (serialization 0.25.5 <-> 0.24.8)
The problem was caused by a conflict between our yaml post processing
(which expects the malformed output of the standard library's yaml
class) and zaml (which produces the right output).

This patch implements the minimal footprint solution.  It just causes
zaml to always escape \r and \n, so that the fixups won't be triggered
and the results will be properly interpreted by any version of puppet.

For Rowlf, this patch should NOT be used, and instead the fix for #3766
(actually rooting out and removing the vestigial fixups) should be used.

Signed-off-by: Markus Roberts <Markus@reality.com>