Adding #2658 - Adding support for run stages
This allows you to specify a run stage for either
a class or a resource.

By default, all classes get directly added to the
'main' stage.  You can create new stages as resources:

    stage { [pre, post]: }

To order stages, use standard relationships:

    stage { pre: before => Stage[main] }

Or use the new relationship syntax:

    stage { pre: } -> Stage[main] -> stage { post: }

Then use the new class parameters to specify a stage:

    class { foo: stage => pre }

If you set a stage on an individual resource, it will
fail; stages can only be set on class resources.

Signed-off-by: Luke Kanies <luke@puppetlabs.com>