Fixing bad test
Actually, just moving the bad around.  The test as written depended on the
implementation details of Parser#file=; when those changed, it broke the test
and this patch fixes the test by updating it's unwarranted assumptions to
match the current details.

The real solution should probably be to ban all such tests and yank them from
the code base without mercy, but that's a task for another day, after we solve
world peace and cure the common cold.