Fixing #3791 - client environment is used
Node#environment wasn't being set correctly, in
that it had to have the right answer out of the gate
or it was never corrected.

It was lazy-binding in 0.25 but I managed to make
it no longer that way.  This resulted in the environment
basically not being set during compilation, so the default
server environment was always used.

Signed-off-by: Luke Kanies <luke@puppetlabs.com>