Fix for conflict between fileserving streams and none-checksums
This was a classic semantic merge conflict; one patch adds a new type of
checksum, with a routine to compute it on files; the other adds streams with
routines to compute all the existing checksum types on them.  They merge
cleanly but theresult is incorrect.

This patch completes the square by adding a none_stream routine to not-compute
a checksum on a stream.