Fixing #2658 - adding backward compatibility for 0.24
The way stages were implemented caused backward compatibility
to be completely broken for 0.24.x.

This commit fixes that, mostly by assuming Stage[main] will be the
top node in the graph rather than Class[main].

Other stages are not supported in 0.24.x, and explicitly throw a warning
(although not an error).

Signed-off-by: Luke Kanies <luke@puppetlabs.com>