Fix #3871 - Add the 'in' operator
This operator allows to find if the left operand is in the right one.
The left operand must be resort to a string, but the right operand can be:
* a string
* an array
* a hash (the search is done on the keys)

This syntax can be used in any place where an expression is supported.

Syntax:
$eatme = 'eat'
if $eatme in ['ate', 'eat'] {
...
}

$value = 'beat generation'
if 'eat' in $value {
  notice("on the road")
}

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>